backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
yfinance==0.2.33
google-generativeai==0.3.1
requests==2.31.0
